This will add the `jsxAttributesIndent` option.
Indenting attributes with a single space provides much more readable code, as the boundaries between attributes, child nodes and closing tags are apparent, and it scales well for large amount of tags:

```html
<div
 attr1="my attr"
 attr2="my other attr">
    <child />
    <child
     nestedAttr2="test"
     nestedAttr3="test" />
</div>
```

The name of the option is not the best though, we can certainly update that.

I got redirected to this project since the idea was not taken all that
heartly in the prettier proper. You can take a look at the discussion
there: https://github.com/prettier/prettier/pull/2082